feat: Re-organize API
Previously, the API was focused on convenience for importing in a
specific way.

This change splits out functionality from convenience.  The convenience
is now done via a prelude module.

Fixes #17
